Added readme.md and helpmarkdown link. Resolved comments
