+++ Tasks/DotNetCoreInstallerV1/Tests/package.json
@@ -22,12 +22,7 @@
     "url": "https://github.com/Microsoft/azure-pipelines-tasks/issues"
   },
   "homepage": "https://github.com/Microsoft/azure-pipelines-tasks#readme",
-  "dependencies": {
-    "@types/node": "^6.0.101",
-    "@types/q": "^1.5.0",
-    "vsts-task-lib": "2.0.6",
-    "vsts-task-tool-lib": "0.4.1",
-    "typed-rest-client": "0.11.0",
-    "semver": "5.6.0"
+  "devDependencies": {
+    "@types/mocha": "^5.2.0"
   }
 }
