+++ Tasks/DotNetCoreInstallerV1/tsconfig.json
@@ -1,8 +1,6 @@
 {
     "compilerOptions": {
         "target": "ES6",
-        "module": "commonjs",
-        "sourceMap": true,
-        "sourceRoot": "C:\\github\\vsts-tasks\\tasks\\dotnetcoreinstaller\\"
+        "module": "commonjs"
     }
 }
