+++ Tasks/DotNetCoreInstallerV1/versionutilities.ts
@@ -2,7 +2,7 @@ import * as path from 'path';
 import * as tl from 'vsts-task-lib/task';
 import * as toolLib from 'vsts-task-tool-lib';
 import * as semver from 'semver';
-import { VersionInfo } from './versionfetcher';
+import { VersionInfo } from "./models"
 
 export function versionCompareFunction(versionA: string, versionB: string): number {
     if (!toolLib.isExplicitVersion(versionA) || !toolLib.isExplicitVersion(versionB)) {
@@ -60,34 +60,6 @@ export function getMatchingVersionFromList(versionInfoList: VersionInfo[], versi
     }
 
     return null;
-}
-
-export class VersionParts {
-    constructor(version: string) {
-        VersionParts.ValidateVersionSpec(version);
-        let parts: string[] = version.split(".");
-
-        this.majorVersion = parts[0];
-        this.minorVersion = parts[1];
-        this.patchVersion = "";
-        if (this.minorVersion != "x") {
-            this.patchVersion = parts[2];
-        }
-    }
-
-    public static ValidateVersionSpec(version): boolean {
-        let parts: string[] = version.split('.');
-        // validate version
-        if (parts.length < 2 || parts.length > 3 || parts[0] == "x" || (parts[1] == "x" && parts.length > 2) || (parts[1] != "x" && parts.length <= 2) || !parts[0] || !parts[1] || (parts.length == 3 && !parts[2]) || Number.isNaN(Number.parseInt(parts[0])) || (Number.isNaN(Number.parseInt(parts[1])) && parts[1] != "x")) {
-            throw tl.loc("VersionNotAllowed", version)
-        }
-
-        return true;
-    }
-
-    majorVersion: string;
-    minorVersion: string;
-    patchVersion: string;
 }
 
 export const Constants = {
